requirements.txt
line-bot-sdk==2.3.0
flask==2.1.2
requests==2.28.0git add .git push
pydantic>=2.0.3,<3
Werkzeug==2.2.2